# Exclude jsdoc pages from bundling and optimization