By excluding the jsdoc output from being handled by assetgraph-builder we gain a significant speed improvement on postprocessing the docs.

Downside: API docs aren't as optimized. But I think the speed increase is worth the trade-off

Relates to #4356


**Latest build on `master`:**
> Build time: 5m 39s. Total deploy time: 5m 41s

**PR preview build from this branch:**
> Build time: 2m 36s. Total deploy time: 2m 38s