@boneskull 2020-07-06
Closing this in lieu of #4358 which just disables optimization entirely for pull requests.